[ELY-137] Move some classes from the org.wildfly.security.auth.provider package to org.wildfly.security.auth.spi and org.wildfly.security.auth.login packages
