+++ src/main/java/org/wildfly/security/auth/provider/SimpleMapBackedSecurityRealm.java
@@ -27,6 +27,11 @@ import java.util.Map;
 import java.util.Set;
 
 import org.wildfly.security.auth.principal.NamePrincipal;
+import org.wildfly.security.auth.spi.AuthenticatedRealmIdentity;
+import org.wildfly.security.auth.spi.CredentialSupport;
+import org.wildfly.security.auth.spi.RealmIdentity;
+import org.wildfly.security.auth.spi.RealmUnavailableException;
+import org.wildfly.security.auth.spi.SecurityRealm;
 import org.wildfly.security.auth.util.NameRewriter;
 import org.wildfly.security.password.Password;
 import org.wildfly.security.password.PasswordFactory;
