+++ src/main/java/org/wildfly/security/auth/spi/CredentialSupport.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.wildfly.security.auth.provider;
+package org.wildfly.security.auth.spi;
 
 /**
  * The level of support for a type of credential.
