+++ src/test/java/org/wildfly/security/auth/KeyStoreBackedSecurityRealmTest.java
@@ -35,10 +35,10 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 import org.wildfly.security.WildFlyElytronProvider;
 import org.wildfly.security.auth.principal.NamePrincipal;
-import org.wildfly.security.auth.provider.CredentialSupport;
 import org.wildfly.security.auth.provider.KeyStoreBackedSecurityRealm;
-import org.wildfly.security.auth.provider.RealmIdentity;
-import org.wildfly.security.auth.provider.SecurityRealm;
+import org.wildfly.security.auth.spi.CredentialSupport;
+import org.wildfly.security.auth.spi.RealmIdentity;
+import org.wildfly.security.auth.spi.SecurityRealm;
 import org.wildfly.security.keystore.EnablingPasswordEntry;
 import org.wildfly.security.password.Password;
 import org.wildfly.security.password.interfaces.BCryptPassword;
